requirements.txt
pandas
numpy
mlflow
scikit-learn
dvc
dagshub
boto3
python-dotenv
pymongo
certifi
requests
pymongo[srv] >= 4.0
-e .